#19 implement component repo, update k8s library versions
This is basically a verbatim copy from the dogu repo.

Regarding the k8s version in the go.mod: importing the component
operator made it necessary to update both component - and dogu operator
because nasty package conflicts (introduced by test mocks on the other
side of our code) could not be properly solve. The commit fix versions
will later be replaced with propver versions.