#19 Fix introduced bug of mishandling wrapped errors in requeue routine
The previously changed if-else block that decides how to requeue on
certain errors did contain a bug and was untested as well. The bug
consisted of only checking the top-level error and not errors inside the
error chain for certain errors.

This commit fixes this bug by switching over the errors (as was done
previously) but then uses a switch case which is a little bit better to
read than if-else laden constructs that might grow additional business
errors in the future.

This commit also tests the above-mentioned behavior.