#19 Revert renaming of generic map function, expand doc comments instead
Map() as a means of value transformation is a language-inspecific
construction, meaning different languages have the same idiom. Thus the
name alone should not differ in order invoke the same idea at the
software developer.

The previous change did change the name and thus the code pattern which
broke with the idiom. This commit reverts this change back to the
original name but expands the doc comments further because Golang
generics make it hard to read the signature in the first place. An
example will provide an idea how to use this function.