#19 Add Component distribution namespace, add doc comments
Previously, the namespace from which the components get distributed (not
to mistake for a k8s namespace) was implicitly assumed to be in the name
of the component or be solely hardcoded as the string "k8s".

This assumption does not hold for the future as there might be different
namespaces where CES components might origin from.

This commit adds