#19 Re-integrate component distribution namespace in Blueprint object
In a previous commit, all component's distribution namespaces did
receive their own field. This is alright for our internal objects. But
the blueprint JSON inside the blueprint CR must not contain this field
as the break would be too strong. Instead the distribution namespace
must be evaluated similar to the dogu namespace.