#19 Embed Component distribution namespace into its name, fix typos
The previous code regarding component diff'ing did end up with a change
to the effective blueprint, aka that what remains of the blueprint and
the blueprint mask.

Anyhow, a Component's distribution namespace did not end up in the
effective blueprint (which is a bug which was not recognized early
enough). This commit joins a Component's distribution namespace in the
same way as it entered the system: in the Component name delimited by a
slash '/'.

The reason for the re-embedding of the distribution namespace
into the name is a possible comparison of a blueprint and its effective
blueprint counter element. A blueprint's effective blueprint should
strongly resemble the original blueprint (exclusive the optional
blueprint mask). Since the original blueprint contains the distribution
namespace within the component's name, the effective blueprint should do
so as well.

Also, this commit fixes some copy-pasta typos from code that was
shamelessly stolen from the dogu counter-part.