#19 Replace Version with masterminds.Semver to support dev versions
Component versions may contain strings like "1.2.3-dev" which is
incompatible with cesapplibs core.Version. This commit replaces the
version library with masterminds.semver/v3. That also means a change in
the empty value of versions which is nil. Absent components do need a
check now so the empty value can be properly detected.