Add example implementation for encrypted config values (WIP)
Encrypted config values for dogus need special handling because the
public key is not available if the dogu is not installed.
So these values for not installed dogus get a special handling.
In such cases the operator will create a kubernetes secret with the
plain config values. The dogu-operator will later fetch those values and
encrypt them in the dogu installation process. #45