#41 use encryption config actions in diff
encrypted config cannot be set directly.
If the dogu exists, we need to encrypt
it ourselves and then persist it.

if the dogu is not installed yet,
we need to persist the sensitive dogu
config in a separate kubernetes secret.
The dogu-operator will encrypt this
config while installing the dogu.

Encrypting it ourselves if the dogu is
not installed is not possible as the
dogu operator creates the encryption key
while installing the dogu. It does not
exist earlier.