#55 Specify attributes of reverse proxy config (wip)
Use specific attributes for the rewrite target, proxy body size and
additional config.

Do not accept values in binary format for the body size because nginx
does not support this.
In the future, we can ignore that if the dogu-operator implements an
abstraction.
Furthermore, this attribute needs to be implemented with a pointer
because "0" is a regular size with the meaning "unlimited".
The operator deletes
the config if the attribut is nil.