#81 update to dogu-op 2.2.0 via hash
The hash is needed as the dogu-op v2 violates
the go standards because it did not contain the
v2 version in the module name. Therefore an
update via 'go mod' or 'go get' is not
possible.

As the new dogu-op version do not contain the
dogu encryption functions anymore, we needed
to strip the code from encryptionAdapter.go
and delete the tests. This is no problem as
this code should be removed anyway because
we do not want to encrypt sensitive dogu config ourselves anymore.