#81 remove doguNotInstalled field in CRD
This field was used to save the diff state
if the config previously existed.
We don't need this field anymore as it was
only necessary to check if we need to
encrypt the config key ourselves or if
we need to write it into a configMap, so
that the dogu operator can encrypt it
after installing the dogu.
As encryption got removed, this info is
not needed anymore.