#93 remove leader-election from singleton operator
Currently, we have exactly one blueprint operator, yet the deployment
uses the `--leader-elect` flag to activate election (among the
astounding number of 1 (one) operator). This commit removes both the
k8s resources and comments the leader-elect flag for consistency
reasons.

These can be restored in an atomic fashion from the git repo if they
are required.