#93 Reduce manager role/binding scope to namespace, remove write access
This commit reduces the manager's (the blueprint operator's, that is)
role from cluster to namespace.

The operator should not have the ability
to create or even delete a blueprint CR that was created by an
administrator (although updating the CR totally belongs into the
operator's scope of work). In this course the YAML files are renamed
accordingly. Furthermore, the *business code* does not issue creation or
deletion of blueprint CRs. Even though the API adapter is fully capable
of doing so, this commit also reduces the manager role's permission
because it that would be currently outside its scope of operation. The
CR should be created and deleted by another entity, the administrator f.
i.