#121 make health checks before and after the blueprint non-blocking
They use a reconcilation backoff now.
Unfortunately, there is one more health
check after applying dogus and before
restarting them.
This one is still necessary as a blocking
health check.

This is because if we want a continuous
process, we will regenerate the state diff
everytime. Therefore, we then have no clue
anymore, if we still need to restart some
dogus, so we need to do this in the same go
as the version change. Problem is, we cannot
restart them immediately because this will
break the creation of service accounts.
Therefore, the health check needs to be in
the same go as well and needs to be blocking
because of that.