#121 WIP: use bluerpint instead of the id to call use cases
this way, whe use cases don't need to load
the blueprint everytime. If there is an
concurrent update, we handle the conflicts
when we try to update the blueprint ourselves.

If we use the same blueprint the whole time
we don't need to recalculate the state diff.
We only have to do this, if we need to
reconcile the CR again, e.g. for
non-blocking health checks.