#121 remove censoring in post-processing
In previous versions, the state diff was
calculated once and then got persisted in
the blueprint CR. Therefore, we had to write
all fields there.

Now we calculate the state diff at every
reconciliation. Because of that we can now
decide which fields we want to publish in the
blueprint CR.