#121 remove StatusPhaseInProgress
That was easier than expected.
The logic to fail if the status is already in progress was completely
in the changeUseCase and therefore very easier to remove.

I removed the whole process of setting the blueprint "in progress" and
there is no replacement via conditions.
The reason for that is, that we want specific conditions for applying
dogus and components separately.