#121 unify health checks, extract component apply
Health Checks
There were problems with the health conditions and events. We always
execute all steps of the blueprint process. Therefore, if a dogu gets
unhealthy after the blueprint completed, the message would be sth like:
"ecosystem unhealthy upfront", which is misleading. This happened with
conditions and events as well.
Another problem was, that we had partial health checks. Setting the
condition after this health checks would lead to a deletion of the
former message which potentially included more info than
e.g. only dogus.
Because of that, we now have only one method for health checks, which
sets the condition and unified health events if something changed. We
now also set the condition to unknown if an error happens while we
fetch the health states. This function gets called multiple times in
the blueprint process.
Maybe we can optimize it further, so we only do a health check in the
same process again if we changed something in the meantime, or we just
reconcile again after applying sth, so that the initial health check
gets called again, but then we need to calculate the state diff again.

Component Apply
Applying dogus and components and waiting for their health was one big
step in the former process. Now health checks and applying components
are their own separate steps.
Previously, the process failed without any retry if the "in progress"
status was and another reconciliation happened. This is not the case
anymore.
The consequence of this is, that the blueprint will not stop applying
if any error happens (unless the error happens everytime).

Removing the "blueprintApplicationFailed", "failed", "completed" and
"blueprintApplied" statuses are the next steps of the refactoring.
We also should move applying dogus to its own use case or rename the
current one.