#121 extract dogu apply
Like components in the last commit, dogus will now be applied in their
own use case separately from health checks etc.
The code for dogus and components is now nearly identical. We don't need
to refactor this duplication as we want to remove components anyway.

The next step will be to remove the dogu restarts as they will be done
by the dogu operator in the future.