#121 remove dogu restarts
Dogu restarts were necessary because they did not use changed config
right away. In the special case of the global config, all dogus needed
to be restarted.

Now, we decided that the dogu operator should restart dogus right away
if config changed. It should not be the responsibility of the
blueprint operator anymore.

To avoid restarting dogus twice if config changes and an upgrade is
needed, the blueprint operator could set a flag in the dogu-CRs, so
that no automatic restart happens. This is an optimization and can
happen later.