#121 rename post-processing to completeBp; remove last status phases
The reason for the post-processing was to deactivate the maintenance
mode and to mark the blueprint as completed or failed.
Now, it would only be called if all other steps before it were
successful, so we not even needed to handle error states anymore.
Therefore, I renamed it to completeBlueprint.

The last status phases could also be removed because with the
refactoring of the apply-process for dogus and components, they aren't
needed any longer.