revert reset dogu version feature
The initial idea was that the blueprint operator resets misconfigured
versions in Dogu-CRs, e.g. a version which does not exist.

Problem with that is, that if the version exists, the dogu operator will
start upgrade scripts. This could lead to schema migrations,
e.g. in postgresql etc. If we then reset the version to the old one,
the dogu could possible not able to start again.
Therefore, we could end up in a problematic timing problem.